Advocating the use of Python's ``with`` keyword for files.
Core Python developers are now pushing "with open(...) as f" as a
best practice and I've found it's something that's not too hard to
introduce to new programmers. It saves them from having to
remember to close files and introduces a concept (context managers)
that's going to be useful when they become more advanced.